refactor: Move DecorateWithEmission to itinerary
