+++ application/src/ext-test/java/org/opentripplanner/ext/emissions/EmissionsTest.java
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner._support.time.ZoneIds;
 import org.opentripplanner.ext.emissions.internal.DefaultEmissionsService;
+import org.opentripplanner.ext.emissions.itinerary.DecorateWithEmission;
 import org.opentripplanner.framework.model.Cost;
 import org.opentripplanner.framework.model.Gram;
 import org.opentripplanner.model.StopTime;
